ProcessProperties to helpers and more comments in LoginScreen and HomeScreen
