+++ src/screens/LoginScreen.tsx
@@ -1,10 +1,11 @@
-import React, { useContext, useState } from 'react';
-import { View, StyleSheet, Alert } from 'react-native';
+import React, { useState } from 'react';
+import { View, StyleSheet } from 'react-native';
 import { Input, Button } from 'react-native-elements';
 import { NativeStackNavigationProp } from '@react-navigation/native-stack';
 import { login } from '../utils/services';
 import { useAuth } from '../hooks/useAuth';
 
+//Defining the types of the nav stack, the LoginScreen props and navigation prop
 type RootStackParamList = {
   Home: undefined;
   Login: undefined;
@@ -19,22 +20,25 @@ type LoginScreenProps = {
 const LoginScreen: React.FC<LoginScreenProps> = ({ navigation }) => {
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
-  //We omit the first argument since we won't be needing it (token)
+  //We omit the first argument since we won't be needing it (token); and we get the setToken funct from useAuth hook
   const {setToken} = useAuth();
+  //I just defined this constant as homming, it's just the device's name, could have been anything
   const deviceName = "homming";
 
+  //Now we'll handle the login process
   const onSubmit = async () => {
     // Perform the login request and store the bearer token
     const data = await login(deviceName, email, password);
     if( data) {
+    //We update the token
     setToken(data.data.plainTextToken)
-    // Save the token securely
-    // Navigate to the Properties screen (Home)
+    // Save the token securely and then Navigate to the Properties screen (Home)
     navigation.navigate('Home');
   }
    
   };
 
+  //Then we render the loginScreen with the Inputs and Button from react-native-elements
   return (
     <View style={styles.container}>
       <Input
